pyproject.toml
[project]
name = "stt4sg-transcribe"
version = "0.1.0"
description = "Speech-to-Text transcription with Faster-Whisper, PyAnnote VAD/diarization, and CTC alignment"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "faster-whisper>=1.2.1",
    "nltk>=3.9.2",
    "numpy>=2.3.5",
    "pandas>=2.3.3",
    "pyannote-audio>=4.0.3",
    "pysubs2>=1.8.0",
    "torch>=2.8.0",
    "torchaudio>=2.8.0",
    "transformers>=4.57.3",
]
